WIP: make pdtr(int, double) be pdtr(double, double)

+++ scipy/special/_cephes.pxd
@@ -68,7 +68,7 @@ cdef extern from "cephes.h" nogil:
     double erf(double x)
     double ndtri(double y0)
     double pdtrc(int k, double m)
-    double pdtr(int k, double m)
+    double pdtr(double k, double m)
     double pdtri(int k, double y)
     double poch(double x, double m)
     double psi(double x)
